Agrega módulo de inventario

+++ models/usuarios.py
@@ -1,8 +1,10 @@
 from sqlalchemy import Column, Integer, String
 from sqlalchemy import ForeignKey
 from sqlalchemy.orm import relationship
+from flask import flash
 from src.models import session, Base
 from flask_login import UserMixin
+from src.models.movimientos import Movimientos
 from werkzeug.security import generate_password_hash, check_password_hash
 import psycopg2
 
@@ -15,10 +17,7 @@ class Usuarios(Base, UserMixin):
     email = Column(String(500), unique=True, nullable=False)
     contraseña = Column(String(30), nullable=False)
     rol = Column(String(30), nullable=False)
-    movimientos = relationship(
-        "Movimientos",
-        back_populates="usuario"
-    )
+    
 
    
 
@@ -39,7 +38,11 @@ class Usuarios(Base, UserMixin):
         return usuario
     
     def eliminar_usuario(id):
-        usuario = session.query(Usuarios).get(id)        
+        usuario = session.query(Usuarios).get(id) 
+        movimientos = session.query(Movimientos).filter_by(usuario_id=id).first()
+        if movimientos:
+            flash("No se puede eliminar el usuario porque tiene movimientos registrados.")
+            return False       
         session.delete(usuario)
         session.commit()
         return usuario
